+++ src/main/scala/nl/biopet/sbtbiopet/BiopetReleaseSettings.scala
@@ -21,17 +21,57 @@
 
 package nl.biopet.sbtbiopet
 
-import nl.biopet.sbtbiopet.BiopetPlugin.autoImport.{
-  biopetIsTool,
-  biopetReleaseInBioconda,
-  biopetReleaseInSonatype
+import com.typesafe.sbt.SbtPgp.autoImportImpl.useGpg
+import nl.biopet.sbtbiopet.BiopetPlugin.autoImport._
+import ohnosequences.sbt.GithubRelease
+import ohnosequences.sbt.GithubRelease.keys._
+import sbt.Keys._
+import sbt.{Def, Opts, Resolver, Setting}
+import sbtassembly.AssemblyPlugin.autoImport.{assembly, assemblyOutputPath}
+import sbtrelease.ReleasePlugin.autoImport.{
+  ReleaseStep,
+  releaseCrossBuild,
+  releaseProcess,
+  releaseStepCommand
 }
-import sbt.Def
-import sbt.Keys.name
-import sbtrelease.ReleasePlugin.autoImport.{ReleaseStep, releaseStepCommand}
 import sbtrelease.ReleaseStateTransformations._
 
 object BiopetReleaseSettings {
+  /*
+   * A sequence of settings specific to release
+   */
+  def biopetReleaseSettings: Seq[Setting[_]] = Seq(
+    resolvers += Resolver.sonatypeRepo("snapshots"),
+    resolvers += Resolver.sonatypeRepo("releases"),
+    releaseCrossBuild := true,
+    publishTo := biopetPublishTo.value,
+    publishMavenStyle := true,
+    useGpg := true,
+    ghreleaseRepoName := biopetUrlName.value,
+    ghreleaseAssets := {
+      val assemblyPath = (assemblyOutputPath in assembly).value
+      if (biopetIsTool.value) Seq(assemblyPath) else Seq()
+    },
+    ghreleaseRepoOrg := githubOrganization.value,
+    //ghreleaseTitle same as upstream default. Specified here to be stable between releases.
+    ghreleaseTitle := { tagName =>
+      s"${name.value} $tagName"
+    },
+    // ghreleaseNotes generic message. (Empty message leads to prompt).
+    ghreleaseNotes := { tagName =>
+      s"Release ${tagName.stripPrefix("v")}"
+    },
+    // ghreleaseGithubToken copied from default for stability.
+    ghreleaseGithubToken := {
+      GithubRelease.defs.githubTokenFromEnv(
+        GithubRelease.defs.defaultTokenEnvVar) orElse
+        GithubRelease.defs.githubTokenFromFile(
+          GithubRelease.defs.defaultTokenFile)
+    },
+    biopetReleaseInBioconda := biopetIsTool.value, // Only release tools in bioconda, not libraries
+    biopetReleaseInSonatype := true,
+    releaseProcess := biopetReleaseProcess.value
+  )
   protected def biopetReleaseStepsStart: Def.Initialize[Seq[ReleaseStep]] = {
     Def.setting[Seq[ReleaseStep]] {
       Seq[ReleaseStep](
@@ -118,4 +158,14 @@ object BiopetReleaseSettings {
         biopetReleaseStepsNextVersion.value
     }
   }
+  /*
+   * Biopet resolver.
+   */
+  protected def biopetPublishTo: Def.Initialize[Option[Resolver]] =
+    Def.setting {
+      if (isSnapshot.value)
+        Some(Opts.resolver.sonatypeSnapshots)
+      else
+        Some(Opts.resolver.sonatypeStaging)
+    }
 }
